endpoints implementiert and all tests pass

+++ src/todo/todo.controller.ts
@@ -1,31 +1,56 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete, UseGuards } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Patch,
+  Param,
+  Delete,
+  UseGuards,
+  NotFoundException,
+  BadRequestException,
+  ForbiddenException,
+} from '@nestjs/common';
 import { TodoService } from './todo.service';
 import { CreateTodoDto } from './dto/create-todo.dto';
 import { UpdateTodoDto } from './dto/update-todo.dto';
 import { JwtAuthGuard } from '../auth/guards/jwt-auth.guard';
 import {
   ApiBearerAuth,
-  ApiCreatedResponse,
+  ApiForbiddenResponse,
   ApiNoContentResponse,
   ApiOkResponse,
   ApiTags,
   ApiUnauthorizedResponse,
 } from '@nestjs/swagger';
-import { ErrorUnauthorizedDto } from '../../old/sample/generic.dtos/error.unauthorized.dto';
+import { ErrorUnauthorizedDto } from '../generic.dtos/error.unauthorized.dto';
 import { ReturnTodoDto } from './dto/return-todo.dto';
+import { UserInfoDto } from '../generic.dtos/userDtoAndEntity';
+import { CurrentUser } from '../decorators/current-user/current-user.decorator';
+import { UserService } from '../auth/user.service/user.service';
 
 @Controller('todo')
 @ApiTags('Methods')
 @ApiBearerAuth()
 export class TodoController {
-  constructor(private readonly todoService: TodoService) {}
+  constructor(
+    private readonly todoService: TodoService,
+    private readonly userService: UserService,
+  ) {}
 
   @UseGuards(JwtAuthGuard)
-  @ApiCreatedResponse({ description: 'created user' })
+  @ApiOkResponse({ description: 'created user' })
   @ApiUnauthorizedResponse({ description: 'Not logged in!', type: ErrorUnauthorizedDto })
   @Post()
   async create(@Body() createTodoDto: CreateTodoDto) {
-    await this.todoService.create(createTodoDto);
+    if (typeof createTodoDto.description !== 'string') {
+      throw new BadRequestException('The required field description is missing in the object!');
+    }
+    if (typeof createTodoDto.title !== 'string') {
+      throw new BadRequestException('The required field title is missing in the object!');
+    }
+
+    return await this.todoService.create(createTodoDto);
   }
 
   @UseGuards(JwtAuthGuard)
@@ -47,7 +72,12 @@ export class TodoController {
   @ApiUnauthorizedResponse({ description: 'Not logged in!', type: ErrorUnauthorizedDto })
   @Get(':id')
   async findOne(@Param('id') id: string) {
-    return await this.todoService.findOne(+id);
+    const result = await this.todoService.findOne(+id);
+
+    if (!result) {
+      throw new NotFoundException(`We did not found a todo item with id ${id}!`);
+    }
+    return result;
   }
 
   @UseGuards(JwtAuthGuard)
@@ -58,14 +88,30 @@ export class TodoController {
   @ApiUnauthorizedResponse({ description: 'Not logged in!', type: ErrorUnauthorizedDto })
   @Patch(':id')
   async update(@Param('id') id: string, @Body() updateTodoDto: UpdateTodoDto) {
+    const result = await this.todoService.findOne(+id);
+    if (!result) {
+      throw new NotFoundException(`We did not found a todo item with id ${id}!`);
+    }
     return await this.todoService.update(+id, updateTodoDto);
   }
 
   @UseGuards(JwtAuthGuard)
-  @ApiNoContentResponse({ description: 'deleted todo' })
+  @ApiOkResponse({ description: 'deleted todo', type: ReturnTodoDto })
   @ApiUnauthorizedResponse({ description: 'Not logged in!', type: ErrorUnauthorizedDto })
+  @ApiForbiddenResponse({ description: 'you need the admin role for this endpoint' })
   @Delete(':id')
-  async remove(@Param('id') id: string) {
-    await this.todoService.remove(+id);
+  async remove(@Param('id') id: string, @CurrentUser() userInfo: UserInfoDto): Promise<ReturnTodoDto> {
+    const user = await this.userService.findOne(userInfo.username);
+    if (user.roles.find((role) => role === 'admin') === undefined) {
+      throw new ForbiddenException('You have to be member of the role admin to call this method!');
+    }
+
+    let result = await this.todoService.findOne(+id);
+    if (!result) {
+      throw new NotFoundException(`We did not found a todo item with id ${id}!`);
+    }
+    result = await this.todoService.remove(+id);
+    console.log(result);
+    return result;
   }
 }
